pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tensorrt-upscaler"
version = "2.0.0"
description = "Minimal dependency TensorRT image upscaler"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["tensorrt", "upscaling", "super-resolution", "image-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: GPU :: NVIDIA CUDA",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

# Core dependencies
dependencies = [
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "PySide6>=6.5.0",
    "cuda-python",
]

[project.optional-dependencies]
# Performance optimizations (highly recommended)
performance = [
    "opencv-python>=4.8.0",
    "numba>=0.58.0",
    "fpng-py>=0.0.4",
]
# All dependencies including performance
full = [
    "opencv-python>=4.8.0",
    "numba>=0.58.0",
    "fpng-py>=0.0.4",
]
# Development tools
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
tensorrt-upscaler = "tensorrt_upscaler.cli:main"

[project.gui-scripts]
tensorrt-upscaler-gui = "tensorrt_upscaler.gui:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tensorrt_upscaler = ["*.ico", "*.png"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
